Move build-scripts/cpe-generate.py to ssg/_build_cpe.py
